AM-1611- moved the deserialize file.
